+++ sdk/tables/Azure.Data.Tables/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 12.0.0-beta.7 (Unreleased)
 
+### Changed
+
+- Removed the `Timestamp` property from the serialized entity when sending it to the service as it is ignored by the service
 
 ## 12.0.0-beta.6 (2021-03-09)